% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_BRR_sampler.R
\name{hmc_sample_BRR}
\alias{hmc_sample_BRR}
\title{HMC sampler for Robust Regression model}
\usage{
hmc_sample_BRR(
  full_data_count,
  C,
  nu,
  sigma,
  prior_means,
  prior_variances,
  iterations,
  warmup,
  chains,
  seed = sample.int(.Machine$integer.max, 1),
  output = F
)
}
\arguments{
\item{full_data_count}{a matrix or dataframe of the unique data with their
corresponding counts}

\item{C}{number of sub-posterior (default to 1)}

\item{nu}{degrees of freedom in t-distribution (must be greater than 2) to
ensure existence of variance}

\item{sigma}{scale parameter in t-distribution}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{iterations}{number of iterations per chain}

\item{warmup}{number of burn in iterations}

\item{chains}{number of chains}

\item{seed}{seed number for random number generation}

\item{output}{boolean value: defaults to T, determines whether or not to
print output to console}
}
\value{
samples from the (sub-)posterior target for the robust regression model
}
\description{
Sample from (sub-)posterior using Stan
}
