% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_BRR_sampler.R
\name{hmc_base_sampler_BRR}
\alias{hmc_base_sampler_BRR}
\title{HMC sampler for base level for robust regression model}
\usage{
hmc_base_sampler_BRR(
  nsamples,
  warmup,
  data_split,
  C,
  nu,
  sigma,
  prior_means,
  prior_variances,
  seed = sample.int(.Machine$integer.max, 1),
  output = F
)
}
\arguments{
\item{nsamples}{number of samples per node}

\item{warmup}{number of burn in iterations}

\item{C}{number of sub-posteriors (default to 1)]}

\item{nu}{degrees of freedom in t-distribution (must be greater than 2) to
ensure existence of variance}

\item{sigma}{scale parameter in t-distribution}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{seed}{seed number for random number generation}

\item{output}{boolean value: defaults to T, determines whether or not
to print output to console}

\item{data}{list of length C where each item is a list where for c=1,...,C,
data_split[\link{c}]$full_data_count is a matrix or data frame of
the unique data with their corresponding counts}
}
\value{
samples from the sub-posterior targets for the split data sets
for the robust regression model
}
\description{
Sample for base level for robust regression
}
